Detect duplicate registrations and report them, add confirmation before removing entrants

+++ OMeetMgmt/remove_from_event.php
@@ -20,6 +20,8 @@ if (!file_exists(get_event_path($event, $key, ".."))) {
   error_and_exit("No such event \"{$event}\", is this an authorized link?\n");
 }
 
+$is_confirmed = isset($_GET["confirmed"]) && ($_GET["confirmed"] == "true");
+
 $results_string = "";
 $competitor_directory = get_competitor_directory($event, $key, "..");
 
@@ -46,14 +48,15 @@ if (!file_exists($removed_competitor_path)) {
 
 $competitor_outputs = array();
 $output_array = array();
+$errors_array = array();
 foreach ($competitor_list as $competitor) {
   if (!is_dir("{$competitor_directory}/{$competitor}")) {
     if (is_dir("{$removed_competitor_path}/{$competitor}")) {
       $removed_competitor_name = file_get_contents("{$removed_competitor_path}/{$competitor}/name");
-      $output_array[] = "{$removed_competitor_name} was already removed.";
+      $errors_array[] = "{$removed_competitor_name} was already removed.";
     }
     else {
-      $output_array[] = "Unknown competitor {$competitor}, manually removed?";
+      $errors_array[] = "Unknown competitor {$competitor}, manually removed?";
     }
   }
   else {
@@ -71,15 +74,19 @@ foreach ($competitor_list as $competitor) {
       foreach ($results_listing as $this_result) {
         if (substr($this_result, 0 - $competitor_id_len - 1) == ",{$competitor}") {
           # remove this entry
-          unlink("{$results_path}/{$this_result}");
+          if ($is_confirmed) {
+            unlink("{$results_path}/{$this_result}");
+          }
           $entry_output .= ", finish marker {$this_result}";
           break;
         }
       }
     }
   
-    $output_array[] = $entry_output;
-    rename("{$competitor_directory}/{$competitor}", "{$removed_competitor_path}/{$competitor}");
+    $output_array[] = array($competitor, $entry_output);
+    if ($is_confirmed) {
+      rename("{$competitor_directory}/{$competitor}", "{$removed_competitor_path}/{$competitor}");
+    }
   }
 }
 
@@ -87,10 +94,33 @@ $event_description = file_get_contents(get_event_path($event, $key, "..") . "/de
 
 echo get_web_page_header(true, true, false);
 
-echo "<p>Removed from {$event_description}\n";
+if ($is_confirmed) {
+  echo "<p>Removed from {$event_description}\n";
+}
+else {
+  echo "<p>Confirm removal from {$event_description}\n";
+  echo "<form action=\"../OMeetMgmt/remove_from_event.php\">\n<input type=hidden name=\"key\" value=\"${key}\">\n";
+  echo "<input type=hidden name=\"event\" value=\"${event}\">\n";
+  echo "<input type=hidden name=\"confirmed\" value=\"true\">\n";
+}
 echo "<ul>\n";
-echo implode("\n", array_map(function ($elt) { return ("<li>{$elt}"); }, $output_array));
+if ($is_confirmed) {
+  echo implode("\n", array_map(function ($elt) { return ("<li>{$elt[1]}"); }, $output_array));
+}
+else {
+  echo implode("\n", array_map(function ($elt) { return ("<li><input type=checkbox name=\"Remove-{$elt[0]}\">{$elt[1]}"); }, $output_array));
+}
 echo "</ul>\n";
+if (!$is_confirmed) {
+  echo "<input type=submit value=\"Confirm removals\">\n";
+  echo "</form>\n";
+}
+
+if (count($errors_array) > 0) {
+  echo "<p><p>Errors found:\n<ul>";
+  echo implode("\n", array_map(function ($elt) { return ("<li>{$elt}"); }, $errors_array));
+  echo "</ul>\n";
+}
 
 echo get_web_page_footer();
 ?>
